pyproject.toml
[project]
name = "agent-reputation-auth"
version = "0.1.0"
description = "SIWE authentication with ERC-8004 agent reputation gating"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "siwe>=4.4.0",
    "agent0-sdk>=0.3rc1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "ruff>=0.8.0",
]
test = [
    "requests>=2.32.0",
    "eth-account>=0.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
